small fixes

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ssuopea <ssuopea@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 12:30:13 by ssuopea           #+#    #+#             */
-/*   Updated: 2025/05/29 15:31:17 by ssuopea          ###   ########.fr       */
+/*   Updated: 2025/05/29 16:56:36 by ssuopea          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,6 @@
 # include <stdarg.h>
 # include <unistd.h>
 
-int	print_base(long long int i, char *base);
+int	print_base(int i, char *base);
 
 #endif
